backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
sqlalchemy==2.0.36
pydantic-settings==2.7.0
openai==1.58.1
langchain==0.3.13
langchain-community==0.3.13
langchain-openai==0.3.1
pypdf==5.1.0
docx2txt==0.8
faiss-cpu==1.9.0
numpy==1.26.4
python-dotenv==1.0.1
python-telegram-bot==21.9
botbuilder-core==4.16.2
botbuilder-integration-aiohttp==4.16.2
httpx==0.28.1
aiofiles==24.1.0
